+++ src/collective/documentgenerator/tests/test_utils.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from collective.documentgenerator.testing import PODTemplateIntegrationTest
-from collective.documentgenerator.utils import update_templates, compute_md5, validate_dict_update
+from collective.documentgenerator.utils import update_templates, compute_md5, update_dict_with_validation
 import collective.documentgenerator as cdg
 from zope.interface import Invalid
 
@@ -44,16 +44,16 @@ class TestUtils(PODTemplateIntegrationTest):
         self.assertEqual(modified, self.test_podtemplate.modification_date)
         self.assertEquals(ret[0][2], 'unchanged')
 
-    def test_validate_dict_update(self):
+    def test_update_dict_with_validation(self):
         # if dict have no common keys, nothing is returned
-        self.assertIsNone(validate_dict_update({}, {}))
-        self.assertIsNone(validate_dict_update({'test1': 1}, {'test2': 2}))
-        self.assertIsNone(validate_dict_update({'test1': 1}, {}))
-        self.assertIsNone(validate_dict_update({}, {'test2': 2}))
-        self.assertIsNone(validate_dict_update({'test': 1}, {'test1': 2, 'test2': 2}))
+        self.assertIsNone(update_dict_with_validation({}, {}))
+        self.assertIsNone(update_dict_with_validation({'test1': 1}, {'test2': 2}))
+        self.assertIsNone(update_dict_with_validation({'test1': 1}, {}))
+        self.assertIsNone(update_dict_with_validation({}, {'test2': 2}))
+        self.assertIsNone(update_dict_with_validation({'test': 1}, {'test1': 2, 'test2': 2}))
 
         # if dict have at least 1 common keys, Invalid is raised
-        self.assertRaises(Invalid, validate_dict_update, {'test': 1, 'test2': 2}, {'test': 2})
-        self.assertRaises(Invalid, validate_dict_update, {'test1': 1, 'test_1': 1}, {'test1': 1, 'test2': 1})
-        self.assertRaises(Invalid, validate_dict_update, {'test': 1}, {'test': 2, 'test1': 1, 'test2': 1})
-        self.assertRaises(Invalid, validate_dict_update, {'test': 1}, {'test': 2})
+        self.assertRaises(Invalid, update_dict_with_validation, {'test': 1, 'test2': 2}, {'test': 2})
+        self.assertRaises(Invalid, update_dict_with_validation, {'test1': 1, 'test_1': 1}, {'test1': 1, 'test2': 1})
+        self.assertRaises(Invalid, update_dict_with_validation, {'test': 1}, {'test': 2, 'test1': 1, 'test2': 1})
+        self.assertRaises(Invalid, update_dict_with_validation, {'test': 1}, {'test': 2})