+++ docs/CHANGES.rst
@@ -13,7 +13,7 @@ Changelog
   [odelaere, sgeulette]
 - Improved validation for ConfigurablePodTemplate
   [odelaere]
-- Added validation to hamper generation context corruption at generation time
+- Added validation to avoid generation context corruption at generation time
   [odelaere]
 
 1.0.5 (2017-03-10)